Move styled components into seperate file
