+++ packages/frontend/src/components/DashboardMyBugsCard.js
@@ -1,26 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { ReactComponent as ReloadIcon } from '../assets/svg/icons/refresh-ccw.svg';
-
-const CardWrapper = styled.article`
-  grid-area: mybugs;
-  display: flex;
-  flex-direction: column;
-  margin-right: 1.5rem;
-  border-radius: 3px;
-  box-shadow: 0 2px 6px 0 rgba(0, 0, 0, 0.2);
-  border-top: 5px solid ${(props) => props.theme.colors.primary};
-  color: #4a4a4a;
-  padding-top: 0.25rem;
-  padding-left: 0.75rem;
-  padding-right: 0.75rem;
-`;
-
-const CardTitle = styled.h3`
-  font-weight: 600;
-  margin-right: 0.75rem;
-  margin-bottom: 0.75rem;
-`;
+import { CardWrapper, CardTitle, CardHeader } from '../ui/components/StyledDashboardCard'
 
 const StyledReloadIcon = styled(ReloadIcon)`
   display: inline-block;
@@ -29,14 +10,8 @@ const StyledReloadIcon = styled(ReloadIcon)`
   height: 16px;
 `;
 
-const CardHeader = styled.div`
-  display: flex;
-  align-items: center;
-  vertical-align: center;
-`;
-
 const BugListTable = styled.table`
-border-collapse: collapse;
+  border-collapse: collapse;
 `;
 
 const BugListTableHeader = styled.th`
@@ -50,10 +25,10 @@ const BugListTableRow = styled.tr`
 `;
 
 const BugListTableLink = styled.a`
-  color: ${props => props.theme.colors.link};
+  color: ${(props) => props.theme.colors.link};
   text-decoration: none;
   &:hover {
-    color: ${props => props.theme.colors.dark};
+    color: ${(props) => props.theme.colors.dark};
     text-decoration: underline;
   }
 `;
@@ -63,7 +38,9 @@ const DashboardMyBugsCard = () => {
   return (
     <CardWrapper>
       <CardHeader>
-        <CardTitle><a href="https://localhost/bugs/user" >My Bugs</a></CardTitle>
+        <CardTitle>
+          <a href="https://localhost/bugs/user">My Bugs</a>
+        </CardTitle>
         <StyledReloadIcon />
       </CardHeader>
       <BugListTable>
@@ -75,10 +52,14 @@ const DashboardMyBugsCard = () => {
         </BugListTableRow>
         <BugListTableRow>
           <td>
-            <BugListTableLink href="https://localhost/#">KEY-1</BugListTableLink>
+            <BugListTableLink href="https://localhost/#">
+              KEY-1
+            </BugListTableLink>
           </td>
           <td>
-            <BugListTableLink href="https://localhost/#">First Example Bug</BugListTableLink>
+            <BugListTableLink href="https://localhost/#">
+              First Example Bug
+            </BugListTableLink>
           </td>
           <td>High</td>
           <td>25/01/2021</td>
